* Rewrite _dot<ZpModule> and _gemv<ZpModule> to chop the input up and pass it along using TypeWrapperRing * Implement _gemm<ZpModule> to chop matrix up and pass it to BLAS * Use float, double rep. types in test-gauss-jordan, test-strassen-winograd * TypeWrapperRing::axpyin now parametrised on output-type * Add previously missing StorageType to ElementVectorTraits, VectorTraits * Implement version of dot taking alternative types for output

+++ tests/test-gauss-jordan.C
@@ -113,7 +113,7 @@ int main (int argc, char **argv)
 
 	parseArguments (argc, argv, args);
 
-	Modular<uint32> GFq (q);
+	Modular<float> GFq (q);
 	GF2 gf2;
 
 	commentator.setBriefReportParameters (Commentator::OUTPUT_CONSOLE, false, false, false);
